GO_CBT_APP_Desktop/requirements.txt
pillow
reportlab
pyinstaller>=6.6
